Move entries from deprecated to legacy
